+++ tests/evaluation/test_runner.py
@@ -2,7 +2,7 @@ from typing import Iterable
 
 import pytest
 
-from intelligence_layer.core import InMemoryTracer
+from intelligence_layer.core import InMemoryTaskSpan, InMemoryTracer
 from intelligence_layer.evaluation import (
     Example,
     InMemoryDatasetRepository,
@@ -40,6 +40,41 @@ def test_runner_runs_dataset(
     failed_runs = list(runner.failed_runs(overview.id, type(None)))
     assert len(failed_runs) == 1
     assert failed_runs[0].example.id == examples[1].id
+
+
+def test_runner_works_without_description(
+    in_memory_dataset_repository: InMemoryDatasetRepository,
+    in_memory_run_repository: InMemoryRunRepository,
+    sequence_examples: Iterable[Example[str, None]],
+) -> None:
+    examples = list(sequence_examples)
+    task = DummyTask()
+    runner = Runner(task, in_memory_dataset_repository, in_memory_run_repository, "")
+
+    dataset_id = in_memory_dataset_repository.create_dataset(
+        examples=examples, dataset_name=""
+    ).id
+    overview = runner.run_dataset(dataset_id)
+    assert overview.description is runner.description
+
+
+def test_runner_has_correct_description(
+    in_memory_dataset_repository: InMemoryDatasetRepository,
+    in_memory_run_repository: InMemoryRunRepository,
+    sequence_examples: Iterable[Example[str, None]],
+) -> None:
+    examples = list(sequence_examples)
+    task = DummyTask()
+    runner = Runner(task, in_memory_dataset_repository, in_memory_run_repository, "foo")
+
+    dataset_id = in_memory_dataset_repository.create_dataset(
+        examples=examples, dataset_name=""
+    ).id
+    run_description = "bar"
+    overview = runner.run_dataset(dataset_id, description=run_description)
+
+    assert runner.description in overview.description
+    assert run_description in overview.description
 
 
 def test_runner_aborts_on_error(
@@ -83,3 +118,7 @@ def test_runner_runs_n_examples(
     assert overview.successful_example_count == 1
     assert overview_with_tracer.successful_example_count == 1
     assert overview_with_tracer.failed_example_count == 0
+
+    entries = tracer.entries
+    assert len(entries) == 1
+    assert all([isinstance(e, InMemoryTaskSpan) for e in entries])